+++ OptiGlue/1.19.3/src/main/kotlin/opekope2/optiglue/Initializer.kt
@@ -3,9 +3,6 @@ package opekope2.optiglue
 import net.fabricmc.fabric.api.resource.ResourceManagerHelper
 import net.minecraft.MinecraftVersion
 import net.minecraft.resource.ResourceType
-import opekope2.optiglue.mc_1_19_3.GlueRegistryLookup
-import opekope2.optiglue.mc_1_19_3.GlueResourceLoader
-import opekope2.optiglue.mc_1_19_3.GlueResourceResolver
 import opekope2.optigui.provider.RegistryLookup
 import opekope2.optigui.provider.ResourceResolver
 import opekope2.optigui.provider.registerProvider
@@ -19,10 +16,10 @@ private val gameVersion = MinecraftVersion.CURRENT.name
 @Suppress("unused")
 fun initialize() {
     // Needed by OptiGUI
-    registerProvider<RegistryLookup>(GlueRegistryLookup())
-    registerProvider<ResourceResolver>(GlueResourceResolver())
+    registerProvider<RegistryLookup>(RegistryLookupImpl())
+    registerProvider<ResourceResolver>(ResourceResolverImpl())
 
-    ResourceManagerHelper.get(ResourceType.CLIENT_RESOURCES).registerReloadListener(GlueResourceLoader)
+    ResourceManagerHelper.get(ResourceType.CLIENT_RESOURCES).registerReloadListener(ResourceLoader)
 
-    logger.info("OptiGlue $modVersion initialized.")
+    logger.info("OptiGlue $modVersion initialized in Minecraft $gameVersion.")
 }
