+++ OptiGlue/1.19.3/src/main/kotlin/opekope2/optiglue/ResourceImpl.kt
@@ -1,4 +1,4 @@
-package opekope2.optiglue.mc_1_19_3
+package opekope2.optiglue
 
 import net.minecraft.resource.ResourceManager
 import net.minecraft.util.Identifier
@@ -7,7 +7,7 @@ import opekope2.optigui.resource.Resource
 import java.util.*
 import kotlin.jvm.optionals.getOrNull
 
-internal class GlueResource(private val manager: ResourceManager, id: Identifier) : Resource(id) {
+internal class ResourceImpl(manager: ResourceManager, id: Identifier) : Resource(id) {
     @OptIn(ExperimentalStdlibApi::class)
     private val resource = manager.getResource(id).getOrNull()
